+++ docker/pip-install.sh
@@ -1,18 +1,14 @@
 #!/bin/bash
 
-# Create and activate Python virtual environment
-pip3 install -U pip
-pip install -U setuptools
-pip install -U virtualenv
-virtualenv -p python3 ENV
+# Create Python virtual environment if not present
+if [ ! -d /opt/openoni/ENV/lib ]; then
+  python3 -m venv ENV
+fi
+
+# Activate the Python virtual environment
 source ENV/bin/activate
 
-# Install / update Open ONI dependencies
-pip install -U -r requirements.pip
-
-# Miscellaneous
-install -d /opt/openoni/static
-install -d /opt/openoni/.python-eggs
-
-# Update requirements.lock
-pip list --format freeze > requirements.lock
+# Install Open ONI dependencies
+# --no-cache-dir disables building local wheels as packages are installed
+# Building wheels provides no benefit when isolated within virtual environment
+pip install --no-cache-dir -r requirements.lock